Update to Python 3.11 and Pipenv

+++ tests/test_feb_stats/test_parsers/test_feb_livescore_stats_transforms.py
@@ -3,7 +3,7 @@ from typing import Any
 
 import pandas as pd
 
-from feb_stats.parsers.feb_stats_transforms import *
+from feb_stats.parsers.feb_livescore_stats_transforms import *
 
 
 class FebStatsTransformsTestCase(unittest.TestCase):
@@ -15,9 +15,7 @@ class FebStatsTransformsTestCase(unittest.TestCase):
         series = pd.Series(["1/4 25%", "6/12 50%"])
         prefix = "x"
         df = transform_cum_stats_shots(series, prefix=prefix)
-        desired_df = pd.DataFrame(
-            {f"{prefix}_made": [1, 6], f"{prefix}_attempted": [4, 12]}, dtype="float32"
-        )
+        desired_df = pd.DataFrame({f"{prefix}_made": [1, 6], f"{prefix}_attempted": [4, 12]}, dtype="float32")
         pd.testing.assert_frame_equal(df, desired_df)
 
     def test_transform_cum_stats_minutes(self) -> None:
@@ -25,9 +23,7 @@ class FebStatsTransformsTestCase(unittest.TestCase):
         series = pd.Series(["26:15", "26:12"])
         prefix = "x"
         df = transform_cum_stats_minutes(series, prefix=prefix)
-        desired_df = pd.DataFrame(
-            {f"{prefix}": ["00:26:15", "00:26:12"]}, dtype="float32"
-        )
+        desired_df = pd.DataFrame({f"{prefix}": ["00:26:15", "00:26:12"]})
         pd.testing.assert_frame_equal(df, desired_df)
 
     def test_transform_cum_stats_blocks(self) -> None:
@@ -35,9 +31,7 @@ class FebStatsTransformsTestCase(unittest.TestCase):
         series = pd.Series(["0 0", "2 5"])
         prefix = "x"
         df = transform_cum_stats_blocks(series, prefix=prefix)
-        desired_df = pd.DataFrame(
-            {f"{prefix}_made": [0, 2], f"{prefix}_received": [0, 5]}, dtype="float32"
-        )
+        desired_df = pd.DataFrame({f"{prefix}_made": [0, 2], f"{prefix}_received": [0, 5]}, dtype="float32")
         pd.testing.assert_frame_equal(df, desired_df)
 
     def test_transform_cum_stats_fouls(self) -> None:
@@ -45,9 +39,7 @@ class FebStatsTransformsTestCase(unittest.TestCase):
         series = pd.Series(["0 0", "2 5"])
         prefix = "x"
         df = transform_cum_stats_fouls(series, prefix=prefix)
-        desired_df = pd.DataFrame(
-            {f"{prefix}_made": [0, 2], f"{prefix}_received": [0, 5]}, dtype="float32"
-        )
+        desired_df = pd.DataFrame({f"{prefix}_made": [0, 2], f"{prefix}_received": [0, 5]}, dtype="float32")
         pd.testing.assert_frame_equal(df, desired_df)
 
     def test_transform_cum_stats_rebounds(self) -> None:
